requirements.txt
SQLAlchemy==1.2.4
xlrd==1.1.0
psycopg2==2.7.3.2
python_ldap==3.0.0b4
progressbar2==3.36.0
